@@ -1,24 +1,40 @@
 package colorizer
 
-import "github.com/logrusorgru/aurora"
+import (
+	"fmt"
 
-// Colorizer returns a colorized severity printer
-type Colorizer struct {
-	Data map[string]string
-}
+	"github.com/logrusorgru/aurora"
+	"github.com/projectdiscovery/gologger"
+	"github.com/projectdiscovery/nuclei/v2/internal/severity"
+)
 
 const (
 	fgOrange uint8 = 208
 )
 
-// New returns a new severity based colorizer
-func New(colorizer aurora.Aurora) *Colorizer {
-	severityMap := map[string]string{
-		"info":     colorizer.Blue("info").String(),
-		"low":      colorizer.Green("low").String(),
-		"medium":   colorizer.Yellow("medium").String(),
-		"high":     colorizer.Index(fgOrange, "high").String(),
-		"critical": colorizer.Red("critical").String(),
+func GetColor(colorizer aurora.Aurora, templateSeverity fmt.Stringer) string {
+	var method func(arg interface{}) aurora.Value
+	switch templateSeverity {
+	case severity.Info:
+		method = colorizer.Blue
+	case severity.Low:
+		method = colorizer.Green
+	case severity.Medium:
+		method = colorizer.Yellow
+	case severity.High:
+		method = func(stringValue interface{}) aurora.Value { return colorizer.Index(fgOrange, stringValue) }
+	case severity.Critical:
+		method = colorizer.Red
+	default:
+		gologger.Warning().Msgf("The '%s' severity does not have an color associated!", templateSeverity)
+		method = colorizer.White
 	}
-	return &Colorizer{Data: severityMap}
+
+	return method(templateSeverity.String()).String()
+}
+
+func New(colorizer aurora.Aurora) func(severity.Severity) string {
+	return func(severity severity.Severity) string {
+		return GetColor(colorizer, severity)
+	}
 }
