@@ -4,7 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"path"
+	"path/filepath"
 	"strings"
 
 	"github.com/projectdiscovery/nuclei/v2/pkg/types"
@@ -26,10 +26,10 @@ func (g *Generator) validate(payloads map[string]interface{}, templatePath strin
 			}
 
 			changed := false
-			pathTokens := strings.Split(templatePath, "/")
+			pathTokens := strings.Split(templatePath, string(os.PathSeparator))
 
 			for i := range pathTokens {
-				tpath := path.Join(strings.Join(pathTokens[:i], "/"), pt)
+				tpath := filepath.Join(filepath.Join(pathTokens[:i]...), pt)
 				if fileExists(tpath) {
 					payloads[name] = tpath
 					changed = true
