@@ -3,8 +3,11 @@ package http
 import (
 	"testing"
 
-	"github.com/projectdiscovery/nuclei/v2/internal/testutils"
 	"github.com/stretchr/testify/require"
+
+	"github.com/projectdiscovery/nuclei/v2/internal/severity"
+	"github.com/projectdiscovery/nuclei/v2/internal/testutils"
+	"github.com/projectdiscovery/nuclei/v2/pkg/model"
 )
 
 func TestHTTPCompile(t *testing.T) {
@@ -14,7 +17,6 @@ func TestHTTPCompile(t *testing.T) {
 	testutils.Init(options)
 	templateID := "testing-http"
 	request := &Request{
-		ID:   templateID,
 		Name: "testing",
 		Payloads: map[string]interface{}{
 			"username": []string{"admin"},
@@ -30,7 +32,7 @@ Accept-Encoding: gzip`},
 	}
 	executerOpts := testutils.NewMockExecuterOptions(options, &testutils.TemplateInfo{
 		ID:   templateID,
-		Info: map[string]interface{}{"severity": "low", "name": "test"},
+		Info: model.Info{SeverityHolder: severity.SeverityHolder{Severity: severity.Low}, Name: "test"},
 	})
 	err := request.Compile(executerOpts)
 	require.Nil(t, err, "could not compile http request")
