@@ -4,14 +4,17 @@ import (
 	"crypto/sha1"
 	"encoding/hex"
 	"os"
-	"path"
+	"path/filepath"
 	"strings"
 	"sync"
 
 	"github.com/owenrumney/go-sarif/sarif"
 	"github.com/pkg/errors"
+
+	"github.com/projectdiscovery/nuclei/v2/internal/severity"
 	"github.com/projectdiscovery/nuclei/v2/pkg/output"
 	"github.com/projectdiscovery/nuclei/v2/pkg/reporting/format"
+	"github.com/projectdiscovery/nuclei/v2/pkg/utils"
 )
 
 // Exporter is an exporter for nuclei sarif output format.
@@ -41,7 +44,7 @@ func New(options *Options) (*Exporter, error) {
 	if err != nil {
 		return nil, errors.Wrap(err, "could not get home dir")
 	}
-	templatePath := path.Join(home, "nuclei-templates")
+	templatePath := filepath.Join(home, "nuclei-templates")
 
 	run := sarif.NewRun("nuclei", "https://github.com/projectdiscovery/nuclei")
 	return &Exporter{options: options, home: templatePath, sarif: report, run: run, mutex: &sync.Mutex{}}, nil
@@ -52,15 +55,15 @@ func (i *Exporter) Export(event *output.ResultEvent) error {
 	templatePath := strings.TrimPrefix(event.TemplatePath, i.home)
 
 	h := sha1.New()
-	h.Write([]byte(event.Host))
+	_, _ = h.Write([]byte(event.Host))
 	templateID := event.TemplateID + "-" + hex.EncodeToString(h.Sum(nil))
 
 	fullDescription := format.MarkdownDescription(event)
 	sarifSeverity := getSarifSeverity(event)
 
 	var ruleName string
-	if s, ok := event.Info["name"]; ok {
-		ruleName = s.(string)
+	if utils.IsNotBlank(event.Info.Name) {
+		ruleName = event.Info.Name
 	}
 
 	var templateURL string
@@ -71,8 +74,8 @@ func (i *Exporter) Export(event *output.ResultEvent) error {
 	}
 
 	var ruleDescription string
-	if d, ok := event.Info["description"]; ok {
-		ruleDescription = d.(string)
+	if utils.IsNotBlank(event.Info.Description) {
+		ruleDescription = event.Info.Description
 	}
 
 	i.mutex.Lock()
@@ -108,17 +111,12 @@ func (i *Exporter) Export(event *output.ResultEvent) error {
 
 // getSarifSeverity returns the sarif severity
 func getSarifSeverity(event *output.ResultEvent) string {
-	var ruleSeverity string
-	if s, ok := event.Info["severity"]; ok {
-		ruleSeverity = s.(string)
-	}
-
-	switch ruleSeverity {
-	case "info":
+	switch event.Info.SeverityHolder.Severity {
+	case severity.Info:
 		return "note"
-	case "low", "medium":
+	case severity.Low, severity.Medium:
 		return "warning"
-	case "high", "critical":
+	case severity.High, severity.Critical:
 		return "error"
 	default:
 		return "note"
